INIT: Simple expressions.
This initial patch features a very simple expression parser, AST, and
stack-based code generator for x86_64, as well as a test suite.

For now, the compiler takes input as a command-line argument and
produces assembly (nasm) on stdout. This is convenient for testing.

The parser produces a simple AST, which provides a visitor interface on
which the code generation is built. Code generation is very naive,
mostly only using two registers and aggressively placing operands on the
stack.