TYPE: Allow assignment of null pointer constant 0 to pointers.
Usually integer literals should not be assignable to a pointer. However,
0 is excepted, as it is a "null pointer constant". In order to
accomplish this, type_is_assignable(Type*, Type*) is augmented with
type_expr_is_assignable(Type*, Expr*), which additionally checks whether
the LHS is a pointer and the RHS is a literal 0.

This is work towards compiling github.com/skeeto/getopt.