LEX: Refactor numeric literal tokenization
This patch slightly cleans up the representation of numeric literal
tokens, pulling out the nested union into its own type, and using a
dedicated enum, rather than a PType (which will be reworked soon).

Support is also added for the z/Z suffix, which is not standard C (but
should be). This was natural to do, since an inherent size_t needs to
exist in order to correctly type sizeof().